requirements.txt
# Dependencies 
requests>=2.31.0        # GitHub API calls
python-dotenv>=1.0.0    # Environment variables
mcp[cli]>=1.4.0         # MCP server functionality
notion-client>=2.3.0    # Notion integration
